refactor: split server snapshot and inspector
